add time intr
