+++ am/arch/mips32-npc/include/npc.h
@@ -1,25 +1,25 @@
 #ifndef __NPC_H__
 #define __NPC_H__
 
-void memory_init();
-void serial_init();
 #define VMEM_ADDR ((void *)0xc0000000)
-#define SCR_WIDTH 640
-#define SCR_HEIGHT 480
+#define SCR_WIDTH 320
+#define SCR_HEIGHT 200
 #define SCR_SIZE (SCR_WIDTH * SCR_HEIGHT)
 
+void memory_init();
+void serial_init();
+void vga_init();
 static inline u8 R(_Pixel p) { return p >> 16; }
 static inline u8 G(_Pixel p) { return p >> 8; }
 static inline u8 B(_Pixel p) { return p; }
-void vga_init();
 static inline u32 Getcolor(_Pixel p){
 	return (R(p) << 8 | G(p) << 4 | B(p) << 2);
 }
-
 static inline _Pixel Getpixel(int color){
 	return color;
 }
-
+u32 GetCount();
+void SetCompare(u32 compare);
 struct TrapFrame{
 	u32 at,
 	v0,v1,
